probe_0002: fixed bug and added debug mode

+++ tests/ttf/probe_0002/test-ctl.py
@@ -11,6 +11,7 @@ import time
 v_target = "none"
 v_operation = "none"
 v_gateware_source = "none"
+v_debug = 0
 
 ########################################################################################################################
 def func_probe():
@@ -31,7 +32,10 @@ def func_probe():
     except (ValueError, KeyError, TypeError):
         print "JSON format error"
     for i in range(len(cmd_list)):
-        subprocess.call(cmd_list[i].split())
+        if v_debug == 0:
+            subprocess.call(cmd_list[i].split())
+        else:
+            print cmd_list[i]
         print "----------------------------------------------------------------------------------------------------"
 
 ########################################################################################################################
@@ -42,25 +46,30 @@ def func_start():
         with open('../devices.json') as json_file:
             data = json.load(json_file)
             for p in data:
+                receivers_string = []
                 receivers = []
                 for q in p['receivers']:
                     if (v_target == str(q['type'])) or (v_target == "all"):
                         relation = "%s:%s" % ((q['dev_name']), (q['slot']))
                         receivers.append(relation)
                         receivers_string = ' '.join(str(x) for x in receivers)
-                        if p['csco_ramdisk'] == "no":
-                            cmd = "timeout 5 ssh %s@%s%s `saftd %s`" % (p['login'], p['name'], p['extension'], receivers_string)
-                        else:
-                            cmd = "timeout 5 ssh %s@%s%s `/usr/sbin/saftd %s`" % (p['login'], p['name'], p['extension'], receivers_string)
-                        cmd_list.append(cmd)
+                if receivers_string:
+                    if p['csco_ramdisk'] == "no":
+                        cmd = "timeout 5 ssh %s@%s%s `saftd %s`" % (p['login'], p['name'], p['extension'], receivers_string)
+                    else:
+                        cmd = "timeout 5 ssh %s@%s%s `/usr/sbin/saftd %s`" % (p['login'], p['name'], p['extension'], receivers_string)
+                    cmd_list.append(cmd)
     except (ValueError, KeyError, TypeError):
         print "JSON format error"
     for i in range(len(cmd_list)):
-        cmd_to_perform = cmd_list[i].split()
-        cmd_to_perform_info = cmd_to_perform[3]
-        print "Starting saftd at %s..." % (cmd_to_perform_info)
-        subprocess.call(cmd_to_perform)
-        time.sleep(1)
+        if v_debug == 0:
+            cmd_to_perform = cmd_list[i].split()
+            cmd_to_perform_info = cmd_to_perform[3]
+            print "Starting saftd at %s..." % (cmd_to_perform_info)
+            subprocess.call(cmd_list[i].split())
+            time.sleep(1)
+        else:
+            print cmd_list[i]
         print "----------------------------------------------------------------------------------------------------"
 
 ########################################################################################################################
@@ -71,17 +80,24 @@ def func_stop():
         with open('../devices.json') as json_file:
             data = json.load(json_file)
             for p in data:
-                if (v_target == str(q['type'])) or (v_target == "all"):
-                    cmd = "timeout 5 ssh %s@%s%s killall saftd" % (p['login'], p['name'], p['extension'])
-                    cmd_list.append(cmd)
+                saftd_stop_found = 0
+                for q in p['receivers']:
+                    if (v_target == str(q['type'])) or (v_target == "all"):
+                        if saftd_stop_found == 0:
+                            cmd = "timeout 5 ssh %s@%s%s killall saftd" % (p['login'], p['name'], p['extension'])
+                            cmd_list.append(cmd)
+                            saftd_stop_found = 1
     except (ValueError, KeyError, TypeError):
         print "JSON format error"
     for i in range(len(cmd_list)):
-        cmd_to_perform = cmd_list[i].split()
-        cmd_to_perform_info = cmd_to_perform[3]
-        print "Stopping saftd at %s..." % (cmd_to_perform_info)
-        subprocess.call(cmd_to_perform)
-        time.sleep(1)
+        if v_debug == 0:
+            cmd_to_perform = cmd_list[i].split()
+            cmd_to_perform_info = cmd_to_perform[3]
+            print "Stopping saftd at %s..." % (cmd_to_perform_info)
+            subprocess.call(cmd_to_perform)
+            time.sleep(1)
+        else:
+            print cmd_list[i]
         print "----------------------------------------------------------------------------------------------------"
 
 ########################################################################################################################
@@ -100,22 +116,27 @@ def func_reset():
         with open('../devices.json') as json_file:
             data = json.load(json_file)
             for p in data:
-                if (v_target == str(q['type'])) or (v_target == "all"):
+                host_reset_found = 0
+                for q in p['receivers']:
                     if (v_target == str(q['type'])) or (v_target == "all"):
-                        for q in p['receivers']:
-                            cmd = "timeout 5 ssh %s@%s%s eb-reset %s" % (p['login'], p['name'], p['extension'], q['slot'])
-                            cmd_list.append(cmd)
-                        if p['reset2host'] == "no":
-                            cmd = "timeout 5 ssh %s@%s%s reboot" % (p['login'], p['name'], p['extension'])
-                            cmd_list.append(cmd)
+                        cmd = "timeout 5 ssh %s@%s%s eb-reset %s" % (p['login'], p['name'], p['extension'], q['slot'])
+                        cmd_list.append(cmd)
+                if p['reset2host'] == "no":
+                    if host_reset_found == 0:
+                        cmd = "timeout 5 ssh %s@%s%s reboot" % (p['login'], p['name'], p['extension'])
+                        cmd_list.append(cmd)
+                        host_reset_found = 1
     except (ValueError, KeyError, TypeError):
         print "JSON format error"
     for i in range(len(cmd_list)):
-        cmd_to_perform = cmd_list[i].split()
-        cmd_to_perform_info = cmd_to_perform[3]
-        print "Resetting device(s) and host at %s..." % (cmd_to_perform_info)
-        subprocess.call(cmd_to_perform)
-        time.sleep(1)
+        if v_debug == 0:
+            cmd_to_perform = cmd_list[i].split()
+            cmd_to_perform_info = cmd_to_perform[3]
+            print "Resetting device(s) and host at %s..." % (cmd_to_perform_info)
+            subprocess.call(cmd_to_perform)
+            time.sleep(1)
+        else:
+            print cmd_list[i]
         print "----------------------------------------------------------------------------------------------------"
 
 ########################################################################################################################
@@ -140,11 +161,14 @@ def func_flash():
     except (ValueError, KeyError, TypeError):
         print "JSON format error"
     for i in range(len(cmd_list)):
-        cmd_to_perform = cmd_list[i].split()
-        cmd_to_perform_info = cmd_to_perform[3]
-        print "Flashing device(s) at %s..." % (cmd_to_perform_info)
-        subprocess.call(cmd_list[i].split())
-        time.sleep(1)
+        if v_debug == 0:
+            cmd_to_perform = cmd_list[i].split()
+            cmd_to_perform_info = cmd_to_perform[3]
+            print "Flashing device(s) at %s..." % (cmd_to_perform_info)
+            subprocess.call(cmd_list[i].split())
+            time.sleep(1)
+        else:
+            print cmd_list[i]
         print "----------------------------------------------------------------------------------------------------"
 
 ########################################################################################################################
@@ -168,7 +192,7 @@ def main():
         else:
             print "Error: Please provide operation name [start target/stop target/restart target/probe target/reset target/{flash target [source URL]}"
             print "Targets: all scu2 scu3 pexarria5 exploder5 microtca pmc vetar2a vetar2a-ee-butis ftm"
-            print "Flash example: flash all http://tsl002.acc.gsi.de/releases/doomsday"
+            print "Flash example: all flash http://tsl002.acc.gsi.de/releases/doomsday"
             exit(1)
     except:
         print "Error: Could not parse given arguments!"
@@ -188,6 +212,7 @@ def main():
     elif v_operation == "flash":
         func_flash()
     else:
+        print "Error: Could not parse given arguments!"
         exit(1)
 
     # Done
